remove RwLock
