--- geyser-plugin-grpc/src/geyser_plugin_grpc.rs
@@ -1,409 +1,0 @@
-use {
-    crate::{
-        accounts_selector::{AccountsSelector, AccountsSelectorConfig},
-        prom::{
-            PrometheusConfig, PrometheusService, BROADCAST_ACCOUNTS_TOTAL, BROADCAST_SLOTS_TOTAL,
-            SLOTS_LAST_PROCESSED,
-        },
-    },
-    geyser_proto::{
-        slot_update::Status as SlotUpdateStatus, update::UpdateOneof, AccountWrite, Ping,
-        SlotUpdate, Update,
-    },
-    log::*,
-    serde::Deserialize,
-    solana_geyser_plugin_interface::geyser_plugin_interface::{
-        GeyserPlugin, GeyserPluginError, ReplicaAccountInfoVersions, Result as PluginResult,
-        SlotStatus,
-    },
-    std::{
-        collections::HashSet,
-        convert::TryInto,
-        fs::read_to_string,
-        sync::{
-            atomic::{AtomicU64, Ordering},
-            Arc, RwLock,
-        },
-    },
-    tokio::{
-        sync::broadcast,
-        time::{sleep, Duration},
-    },
-    tonic::transport::Server,
-};
-
-pub mod geyser_proto {
-    tonic::include_proto!("accountsdb");
-}
-
-pub mod geyser_service {
-    use {
-        super::geyser_proto::{
-            accounts_db_server::AccountsDb, update::UpdateOneof, SubscribeRequest,
-            SubscribeResponse, Update, UpdateAccountsSelectorRequest,
-            UpdateAccountsSelectorResponse,
-        },
-        crate::accounts_selector::{AccountsSelector, AccountsSelectorConfig},
-        log::*,
-        serde::Deserialize,
-        std::sync::{
-            atomic::{AtomicU64, Ordering},
-            Arc, RwLock,
-        },
-        tokio::sync::{broadcast, mpsc},
-        tokio_stream::wrappers::ReceiverStream,
-        tonic::{Code, Request, Response, Status},
-    };
-
-    #[derive(Clone, Debug, Deserialize)]
-    pub struct ServiceConfig {
-        broadcast_buffer_size: usize,
-        subscriber_buffer_size: usize,
-    }
-
-    #[derive(Debug)]
-    pub struct Service {
-        pub sender: broadcast::Sender<Update>,
-        pub config: ServiceConfig,
-        pub highest_write_slot: Arc<AtomicU64>,
-        pub accounts_selector: Arc<RwLock<AccountsSelector>>,
-    }
-
-    impl Service {
-        pub fn new(
-            config: ServiceConfig,
-            highest_write_slot: Arc<AtomicU64>,
-            accounts_selector: Arc<RwLock<AccountsSelector>>,
-        ) -> Self {
-            let (tx, _) = broadcast::channel(config.broadcast_buffer_size);
-            Self {
-                sender: tx,
-                config,
-                highest_write_slot,
-                accounts_selector,
-            }
-        }
-    }
-
-    #[tonic::async_trait]
-    impl AccountsDb for Service {
-        type SubscribeStream = ReceiverStream<Result<Update, Status>>;
-
-        async fn subscribe(
-            &self,
-            _request: Request<SubscribeRequest>,
-        ) -> Result<Response<Self::SubscribeStream>, Status> {
-            info!("new subscriber");
-            let (tx, rx) = mpsc::channel(self.config.subscriber_buffer_size);
-            let mut broadcast_rx = self.sender.subscribe();
-
-            tx.send(Ok(Update {
-                update_oneof: Some(UpdateOneof::SubscribeResponse(SubscribeResponse {
-                    highest_write_slot: self.highest_write_slot.load(Ordering::SeqCst),
-                })),
-            }))
-            .await
-            .unwrap();
-
-            tokio::spawn(async move {
-                let mut exit = false;
-                while !exit {
-                    let fwd = broadcast_rx.recv().await.map_err(|err| {
-                        // Note: If we can't keep up pulling from the broadcast
-                        // channel here, there'll be a Lagged error, and we'll
-                        // close the connection because data was lost.
-                        warn!("error while receiving message to be broadcast: {:?}", err);
-                        exit = true;
-                        Status::new(Code::Internal, err.to_string())
-                    });
-                    if let Err(_err) = tx.send(fwd).await {
-                        info!("subscriber stream closed");
-                        exit = true;
-                    }
-                }
-            });
-            Ok(Response::new(ReceiverStream::new(rx)))
-        }
-
-        async fn update_accounts_selector(
-            &self,
-            request: Request<UpdateAccountsSelectorRequest>,
-        ) -> Result<Response<UpdateAccountsSelectorResponse>, Status> {
-            let (is_ok, error_message) =
-                match serde_json::from_str::<AccountsSelectorConfig>(&request.get_ref().config)
-                    .map_err(|error| error.to_string())
-                    .and_then(|config| {
-                        AccountsSelector::from_config(&config).map_err(|error| error.to_string())
-                    }) {
-                    Ok(accounts_selector_new) => {
-                        *self.accounts_selector.write().unwrap() = accounts_selector_new;
-                        (true, String::new())
-                    }
-                    Err(error) => (false, error),
-                };
-
-            Ok(Response::new(UpdateAccountsSelectorResponse {
-                is_ok,
-                error_message,
-            }))
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct PluginConfig {
-    pub bind_address: String,
-    pub accounts_selector: AccountsSelectorConfig,
-    pub service_config: geyser_service::ServiceConfig,
-    pub zstd_compression: bool,
-    #[serde(default)]
-    pub prometheus: Option<PrometheusConfig>,
-}
-
-#[derive(Debug)]
-pub struct PluginInner {
-    runtime: tokio::runtime::Runtime,
-    prometheus: PrometheusService,
-    server_broadcast: broadcast::Sender<Update>,
-    server_exit_sender: broadcast::Sender<()>,
-    accounts_selector: Arc<RwLock<AccountsSelector>>,
-
-    /// Largest slot that an account write was processed for
-    highest_write_slot: Arc<AtomicU64>,
-
-    /// Accounts that saw account writes
-    ///
-    /// Needed to catch writes that signal account closure, where
-    /// lamports=0 and owner=system-program.
-    active_accounts: RwLock<HashSet<[u8; 32]>>,
-    zstd_compression: bool,
-}
-
-impl PluginInner {
-    fn broadcast(&self, update: UpdateOneof) {
-        // Don't care about the error that happens when there are no receivers.
-        let _ = self.server_broadcast.send(Update {
-            update_oneof: Some(update),
-        });
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct Plugin {
-    data: Option<PluginInner>,
-}
-
-impl GeyserPlugin for Plugin {
-    fn name(&self) -> &'static str {
-        "GeyserPluginGrpc"
-    }
-
-    fn on_load(&mut self, config_file: &str) -> PluginResult<()> {
-        solana_logger::setup_with_default("info");
-        info!(
-            "Loading plugin {:?} from config_file {:?}",
-            self.name(),
-            config_file
-        );
-
-        let config = read_to_string(config_file).map_err(GeyserPluginError::ConfigFileOpenError)?;
-        let config: PluginConfig = serde_json::from_str(&config).map_err(|error| {
-            GeyserPluginError::ConfigFileReadError {
-                msg: format!("Failed to read config from the file: {:?}", error),
-            }
-        })?;
-        let accounts_selector = Arc::new(RwLock::new(
-            AccountsSelector::from_config(&config.accounts_selector).unwrap(),
-        ));
-
-        let runtime = tokio::runtime::Runtime::new().unwrap();
-        let prometheus = PrometheusService::new(&runtime, config.prometheus);
-
-        let bind_address =
-            config
-                .bind_address
-                .parse()
-                .map_err(|err| GeyserPluginError::ConfigFileReadError {
-                    msg: format!("Error parsing the bind_address {:?}", err),
-                })?;
-
-        let highest_write_slot = Arc::new(AtomicU64::new(0));
-        let service = geyser_service::Service::new(
-            config.service_config,
-            highest_write_slot.clone(),
-            Arc::clone(&accounts_selector),
-        );
-
-        let (server_exit_sender, mut server_exit_receiver) = broadcast::channel::<()>(1);
-        let server_broadcast = service.sender.clone();
-
-        let server = geyser_proto::accounts_db_server::AccountsDbServer::new(service)
-            .accept_gzip()
-            .send_gzip();
-        runtime.spawn(Server::builder().add_service(server).serve_with_shutdown(
-            bind_address,
-            async move {
-                let _ = server_exit_receiver.recv().await;
-            },
-        ));
-        let server_broadcast_c = server_broadcast.clone();
-        let mut server_exit_receiver = server_exit_sender.subscribe();
-        runtime.spawn(async move {
-            loop {
-                // Don't care about the error if there are no receivers.
-                let _ = server_broadcast_c.send(Update {
-                    update_oneof: Some(UpdateOneof::Ping(Ping {})),
-                });
-
-                tokio::select! {
-                    _ = server_exit_receiver.recv() => { break; },
-                    _ = sleep(Duration::from_secs(5)) => {},
-                }
-            }
-        });
-
-        self.data = Some(PluginInner {
-            runtime,
-            prometheus,
-            server_broadcast,
-            server_exit_sender,
-            accounts_selector,
-            highest_write_slot,
-            active_accounts: RwLock::new(HashSet::new()),
-            zstd_compression: config.zstd_compression,
-        });
-
-        Ok(())
-    }
-
-    fn on_unload(&mut self) {
-        info!("Unloading plugin: {:?}", self.name());
-
-        let data = self.data.take().expect("plugin must be initialized");
-
-        data.prometheus.shutdown();
-        let _ = data.server_exit_sender.send(());
-        data.runtime.shutdown_background();
-    }
-
-    fn update_account(
-        &mut self,
-        account: ReplicaAccountInfoVersions,
-        slot: u64,
-        is_startup: bool,
-    ) -> PluginResult<()> {
-        let data = self.data.as_ref().expect("plugin must be initialized");
-        match account {
-            ReplicaAccountInfoVersions::V0_0_1(account) => {
-                if account.pubkey.len() != 32 {
-                    error!(
-                        "bad account pubkey length: {}",
-                        bs58::encode(account.pubkey).into_string()
-                    );
-                    return Ok(());
-                }
-
-                // Select only accounts configured to look at, plus writes to accounts
-                // that were previously selected (to catch closures and account reuse)
-                let is_selected = data
-                    .accounts_selector
-                    .read()
-                    .unwrap()
-                    .is_account_selected(account.pubkey, account.owner);
-                let previously_selected = {
-                    let read = data.active_accounts.read().unwrap();
-                    read.contains(&account.pubkey[0..32])
-                };
-                if !is_selected && !previously_selected {
-                    return Ok(());
-                }
-
-                // If the account is newly selected, add it
-                if !previously_selected {
-                    let mut write = data.active_accounts.write().unwrap();
-                    write.insert(account.pubkey.try_into().unwrap());
-                }
-
-                data.highest_write_slot.fetch_max(slot, Ordering::SeqCst);
-
-                debug!(
-                    "Updating account {:?} with owner {:?} at slot {:?}",
-                    bs58::encode(account.pubkey).into_string(),
-                    bs58::encode(account.owner).into_string(),
-                    slot,
-                );
-
-                // zstd compress if enabled
-                let account_data = if data.zstd_compression {
-                    match zstd::encode_all(account.data, 0) {
-                        Ok(vec) => vec,
-                        Err(error) => {
-                            error!("zstd_decompress compression failed: {:?}", error);
-                            account.data.to_vec()
-                        }
-                    }
-                } else {
-                    account.data.to_vec()
-                };
-
-                data.broadcast(UpdateOneof::AccountWrite(AccountWrite {
-                    slot,
-                    is_startup,
-                    write_version: account.write_version,
-                    pubkey: account.pubkey.to_vec(),
-                    lamports: account.lamports,
-                    owner: account.owner.to_vec(),
-                    executable: account.executable,
-                    rent_epoch: account.rent_epoch,
-                    data: account_data,
-                    is_selected,
-                }));
-
-                BROADCAST_ACCOUNTS_TOTAL.inc();
-            }
-        }
-
-        Ok(())
-    }
-
-    fn update_slot_status(
-        &mut self,
-        slot: u64,
-        parent: Option<u64>,
-        status: SlotStatus,
-    ) -> PluginResult<()> {
-        let data = self.data.as_ref().expect("plugin must be initialized");
-        debug!("Updating slot {:?} at with status {:?}", slot, status);
-
-        let (status, label) = match status {
-            SlotStatus::Processed => (SlotUpdateStatus::Processed, "processed"),
-            SlotStatus::Confirmed => (SlotUpdateStatus::Confirmed, "confirmed"),
-            SlotStatus::Rooted => (SlotUpdateStatus::Rooted, "rooted"),
-        };
-
-        data.broadcast(UpdateOneof::SlotUpdate(SlotUpdate {
-            slot,
-            parent,
-            status: status as i32,
-        }));
-
-        SLOTS_LAST_PROCESSED
-            .with_label_values(&[label])
-            .set(slot as i64);
-        BROADCAST_SLOTS_TOTAL.with_label_values(&[label]).inc();
-
-        Ok(())
-    }
-}
-
-#[no_mangle]
-#[allow(improper_ctypes_definitions)]
-/// # Safety
-///
-/// This function returns the Plugin pointer as trait GeyserPlugin.
-pub unsafe extern "C" fn _create_plugin() -> *mut dyn GeyserPlugin {
-    let plugin = Plugin::default();
-    let plugin: Box<dyn GeyserPlugin> = Box::new(plugin);
-    Box::into_raw(plugin)
-}
