+++ geyser-plugin-grpc/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod accounts_selector;
-pub mod geyser_plugin_grpc;
+pub mod plugin;
 pub mod prom;
+pub mod service;
 pub mod version;
